+++ zio-flow-server/src/main/scala/zio/flow/server/FlowTemplates.scala
@@ -1,7 +1,8 @@
 package zio.flow.server
 
 import zio._
-import zio.flow.internal.{KeyValueStore, Namespaces, Timestamp}
+import zio.flow.internal.Namespaces
+import zio.flow.runtime.{KeyValueStore, Timestamp}
 import zio.flow.{FlowId, TemplateId, ZFlowExecutor, ZFlowTemplate}
 import zio.schema.codec.JsonCodec
 import zio.stream.ZStream
