+++ zio-flow/shared/src/main/scala/zio/flow/ExecutingFlow.scala
@@ -16,7 +16,7 @@
 
 package zio.flow
 
-import zio.flow.internal.DurablePromise
+import zio.flow.runtime.DurablePromise
 import zio.schema.{Schema, TypeId}
 
 final case class ExecutingFlow[+E, +A](id: FlowId, result: DurablePromise[_, _])
