+++ zio-flow/shared/src/main/scala/zio/flow/RemoteContext.scala
@@ -18,6 +18,7 @@ package zio.flow
 
 import zio.ZIO
 import zio.flow.internal._
+import zio.flow.runtime.{DurableLog, ExecutorError, Timestamp}
 import zio.schema.{DynamicValue, Schema}
 import zio.stm.TMap
