+++ zio-flow/shared/src/main/scala/zio/flow/ZFlowExecutor.scala
@@ -18,15 +18,8 @@ package zio.flow
 
 import zio._
 import zio.flow.internal.PersistentExecutor.FlowResult
-import zio.flow.internal.{
-  DefaultOperationExecutor,
-  DurableLog,
-  DurablePromise,
-  ExecutorError,
-  IndexedStore,
-  KeyValueStore,
-  PersistentExecutor
-}
+import zio.flow.internal.{DefaultOperationExecutor, PersistentExecutor}
+import zio.flow.runtime.{DurableLog, DurablePromise, ExecutorError, IndexedStore, KeyValueStore}
 import zio.flow.serialization.{Deserializer, Serializer}
 import zio.schema.{DynamicValue, Schema}
