+++ zio-flow/shared/src/main/scala/zio/flow/internal/PersistentExecutor.scala
@@ -18,13 +18,13 @@ package zio.flow.internal
 
 import zio._
 import zio.flow.Remote.UnboundRemoteFunction
-import zio.flow.internal.IndexedStore.Index
-import zio.flow.metrics
+import zio.flow.runtime.IndexedStore.Index
 import zio.flow.remote.DynamicValueHelpers
 import zio.flow.serialization._
 import zio.flow._
 import zio.flow.internal.PersistentExecutor.GarbageCollectionCommand
-import zio.flow.metrics.{TransactionOutcome, finishedFlowAge, finishedFlowCount, flowTotalExecutionTime}
+import zio.flow.runtime.metrics.{TransactionOutcome, finishedFlowAge, finishedFlowCount, flowTotalExecutionTime}
+import zio.flow.runtime.{DurableLog, DurablePromise, ExecutorError, KeyValueStore, Timestamp, metrics}
 import zio.schema.{CaseSet, DeriveSchema, DynamicValue, Schema, TypeId}
 
 import java.nio.charset.StandardCharsets
