+++ zio-flow/shared/src/main/scala/zio/flow/runtime/KeyValueStore.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package zio.flow.internal
+package zio.flow.runtime
 
 import zio.stream.ZStream
 import zio.{Chunk, IO, Ref, ZIO, ZLayer}
