+++ zio-flow/shared/src/main/scala/zio/flow/runtime/metrics/VariableKind.scala
@@ -14,18 +14,21 @@
  * limitations under the License.
  */
 
-package zio.flow.metrics
+package zio.flow.runtime.metrics
 
-sealed trait StartType
+sealed trait VariableKind
 
-object StartType {
-  case object Fresh extends StartType
+object VariableKind {
+  case object TopLevel extends VariableKind
 
-  case object Continued extends StartType
+  case object Forked extends VariableKind
 
-  def toLabel(startType: StartType): String =
-    startType match {
-      case StartType.Fresh     => "fresh"
-      case StartType.Continued => "continued"
+  case object Transactional extends VariableKind
+
+  def toLabel(kind: VariableKind): String =
+    kind match {
+      case VariableKind.TopLevel      => "toplevel"
+      case VariableKind.Forked        => "forked"
+      case VariableKind.Transactional => "transactional"
     }
 }
