+++ zio-flow/shared/src/test/scala/zio/flow/utils/ZFlowAssertionSyntax.scala
@@ -18,8 +18,8 @@ package zio.flow.utils
 
 import zio._
 import zio.flow.internal.PersistentExecutor.FlowResult
-import zio.flow.internal.{DurableLog, ExecutorError, KeyValueStore}
 import zio.flow.mock.MockedOperation
+import zio.flow.runtime.{DurableLog, ExecutorError, KeyValueStore}
 import zio.flow.{Configuration, FlowId, ZFlow, ZFlowExecutor}
 import zio.schema.{DynamicValue, Schema}
